# openapi_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "bada-openstack-api",
    "version": "0.1.0",
    "description": "bada-openstack-api document"
  },
  "paths": {
...
    "/api/auth/login/": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "login 성공시 토큰을 쿠키에 발급한다\n\n- 200\n\n- 400 : 필드 오류\n\n- 401 : 로그인 실패",
        "operationId": "login_api_auth_login__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": "로그인 성공"
              }
            }
          },
          "400": {
            "description": "필드 조건 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PydanticException"
                }
              }
            }
          },
          "401": {
            "description": "로그인 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicException"
                },
                "example": {
                  "error_type": "error",
                  "message": "The request you have made requires authentication.",
                  "detail": ""
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout/": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "description": "logout 성공시 쿠키에 저장된 토큰을 제거한다",
        "operationId": "logout_api_auth_logout__post",
        "responses": {
          "200": {
            "description": "로그아웃 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": "로그아웃 성공"
              }
            }
          },
          "401": {
            "description": "로그아웃 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicException"
                },
                "example": {
                  "error_type": "error",
                  "message": "요청 헤더에 토큰이 존재하지 않습니다",
                  "detail": ""
                }
              }
            }
          }
        }
      }
    },
...
    "/api/floatingips/": {
      "get": {
        "tags": [
          "floatingip"
        ],
        "summary": "get list of floating ips",
        "description": "조건에 해당하는 floating ip list를 반환한다\n\n\n- 200\n- 400",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "요청 페이지"
          },
          {
            "in": "query",
            "name": "per_page",
            "schema": {
              "type": "integer"
            },
            "description": "한 페이지당 갯수"
          },
          {
            "in": "query",
            "name": "sort_by",
            "schema": {
              "type": "string"
            },
            "description": "정렬 필드명"
          },
          {
            "in": "query",
            "name": "order_by",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "정렬 조건"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FloatingipResponse"
                  }
                },
                "example": [
                  {
                    "floatingip_id": "c668e19b-14af-4c11-bb47-7ae86fbae17e",
                    "ip_address": "210.111.123.142",
                    "description": "test floating ip description",
                    "project_id": "e25641b6-f60a-4cab-9da8-a9729fdb66df",
                    "status": "Active",
                    "server": {
                      "server_id": "34b835f0-9443-4d37-b08b-efca349f7716",
                      "name": "test_server"
                    },
                    "created_at": "2024-02-20T05:42:13.618504",
                    "updated_at": "2024-02-20T05:42:13.618504",
                    "deleted_at": null
                  },
                  {
                    "floatingip_id": "c668e19b-14af-4c11-bb47-7ae86fbae17e",
                    "ip_address": "210.111.123.143",
                    "description": "test floating ip description",
                    "project_id": "e25641b6-f60a-4cab-9da8-a9729fdb66df",
                    "status": "Active",
                    "server": {
                      "server_id": "34b835f0-9443-4d37-b08b-efca349f7716",
                      "name": "test_server"
                    },
                    "created_at": "2024-02-20T05:42:13.618504",
                    "updated_at": "2024-02-20T05:42:13.618504",
                    "deleted_at": null
                  }
                ]
              }
            }
          },
          "400": {
            "description": "필드 조건 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PydanticException"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "floatingip"
        ],
        "summary": "create floating ip",
        "description": "\npublic network로부터 floating ip를 생성한다. \n\n- 201\n\n- 400 : 필드 조건 오류 (description.length > 255)\n\n- 404 : 해당 server 존재하지 않음\n\n- 409 : floating ip quota 초과, 이미 해당 서버에 floating ip 할당되어 있는 경우",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FloatingipCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloatingipResponse"
                },
                "example": {
                  "floatingip_id": "c668e19b-14af-4c11-bb47-7ae86fbae17e",
                  "ip_address": "210.111.123.142",
                  "description": "test floating ip description",
                  "project_id": "e25641b6-f60a-4cab-9da8-a9729fdb66df",
                  "status": "Active",
                  "server": {
                    "server_id": "34b835f0-9443-4d37-b08b-efca349f7716",
                    "name": "test_server"
                  },
                  "created_at": "2024-02-20T05:42:13.618504",
                  "updated_at": "2024-02-20T05:42:13.618504",
                  "deleted_at": null
                }
              }
            }
          },
          "400": {
            "description": "필드 조건 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PydanticException"
                }
              }
            }
          },
          "404": {
            "description": "port not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicException"
                }
              }
            }
          },
          "409": {
            "description": "잔여 floating ip 없음 / 이미 해당 서버는 연결되어 있음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicException"
                }
              }
            }
          }
        }
      }
    },
    "/api/floatingips/{floatingip_id}/": {
      "parameters": [
        {
          "in": "path",
          "name": "floatingip_id",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "floatingip uuid"
        }
      ],
      "get": {
        "tags": [
          "floatingip"
        ],
        "summary": "get specific floating ip",
        "description": "해당하는 uuid의 floating ip를 반환한다\n\n- 200\n\n- 404 : 해당 floating ip가 존재하지 않음",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloatingipResponse"
                },
                "example": {
                  "floatingip_id": "c668e19b-14af-4c11-bb47-7ae86fbae17e",
                  "ip_address": "210.111.123.142",
                  "description": "test floating ip description",
                  "project_id": "e25641b6-f60a-4cab-9da8-a9729fdb66df",
                  "status": "Active",
                  "server": {
                    "server_id": "34b835f0-9443-4d37-b08b-efca349f7716",
                    "name": "test_server"
                  },
                  "created_at": "2024-02-20T05:42:13.618504",
                  "updated_at": "2024-02-20T05:42:13.618504",
                  "deleted_at": null
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicException"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "floatingip"
        ],
        "summary": "update specific floating ip's info (description)",
        "description": "\n해당하는 uuid의 floating ip의 정보(description)를 수정한다\n\n- 200\n\n- 400 : 필드 조건 오류\n\n- 404 : 해당 floating ip가 존재하지 않음",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FloatingipUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloatingipResponse"
                },
                "example": {
                  "floatingip_id": "c668e19b-14af-4c11-bb47-7ae86fbae17e",
                  "ip_address": "210.111.123.142",
                  "description": "test floating ip description",
                  "project_id": "e25641b6-f60a-4cab-9da8-a9729fdb66df",
                  "status": "Active",
                  "server": {
                    "server_id": "34b835f0-9443-4d37-b08b-efca349f7716",
                    "name": "test_server"
                  },
                  "created_at": "2024-02-20T05:42:13.618504",
                  "updated_at": "2024-02-20T05:42:13.618504",
                  "deleted_at": null
                }
              }
            }
          },
          "400": {
            "description": "필드 조건 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PydanticException"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicException"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "floatingip"
        ],
        "summary": "delete specific floating ip",
        "description": "해당하는 uuid의 floating ip를 삭제한다. 서버에 할당한 경우 삭제할 수 없다.\n\n- 204\n\n- 404 : 해당 floating ip가 존재하지 않음\n\n- 409 : 연결된 서버 있는 경우, 이미 floatingip 삭제된 경우",
        "responses": {
          "204": {
            "description": "No-Content"
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicException"
                }
              }
            }
          },
          "409": {
            "description": "conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicException"
                }
              }
            }
          }
        }
      }
    },
    "/api/floatingips/{floatingip_id}/ports/": {
      "patch": {
        "tags": [
          "floatingip"
        ],
        "summary": "attach/detach to port",
        "description": "**need polling**\n\n해당하는 floatingip와 port를 연결하거나 해제(port_id == null)한다\n\n\n- 202\n\n- 400 : 필드 조건 오류\n\n- 404 : 해당하는 floatingip 없는 경우, portid 없는 경우\n\n- 409: 이미 floatingip 삭제된 경우, 해당 서버 불가능한 경우",
        "parameters": [
          {
            "in": "path",
            "name": "floatingip_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FloatingipUpdatePortRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "success + status 추적 task실행",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloatingipResponse"
                }
              }
            }
          },
          "400": {
            "description": "필드 조건 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PydanticException"
                }
              }
            }
          },
          "404": {
            "description": "해당 서버를 찾을 수 없음. 해당 floatingip 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicException"
                }
              }
            }
          },
          "409": {
            "description": "conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicException"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TokenCreateRequest": {
        "title": "TokenCreateRequest",
        "description": "사용자가 보낼 login request",
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "PydanticException": {
        "properties": {
          "error_type": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "properties": {
              "input": {
                "properties": {
                  "field_name": {
                    "type": "string"
                  }
                },
                "type": "object"
              },
              "loc": {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    }
                  ]
                },
                "type": "array"
              }
            },
            "type": "object"
          }
        },
        "type": "object",
        "title": "PydanticException",
        "description": "필드 조건 오류"
      },
      "BasicException": {
        "properties": {
          "error_type": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detail": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object"
              }
            ]
          }
        }
      },
...
      "FloatingipResponse": {
        "properties": {
          "floatingip_id": {
            "type": "string"
          },
          "ip_address": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "server": {
            "$ref": "#/components/schemas/ServerOverallResponse"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "deleted_at": {
            "type": "string"
          }
        }
      },
      "FloatingipCreateRequest": {
        "properties": {
          "description": {
            "type": "string"
          }
        }
      },
      "FloatingipUpdateRequest": {
        "properties": {
          "description": {
            "type": "string"
          }
        }
      },
      "FloatingipUpdatePortRequest": {
        "properties": {
          "port_id": {
            "type": "string"
          }
        }
      },
...
      "ServerOverallResponse": {
        "description": "연관된 서버 보여줄 때",
        "properties": {
          "server_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      }
    }
  }
}